feat(home): add wish products section
